fix(monorepo): Fix Eslint issues + clean code
