+++ apps/frontend/src/composables/useConnection.ts
@@ -1,4 +1,3 @@
-// src/composables/useConnection.ts
 import { invoke } from "@tauri-apps/api/core";
 import { ref } from "vue";
