+++ apps/frontend/src/composables/useStreaming.ts
@@ -1,4 +1,3 @@
-// src/composables/useStreaming.ts
 import { invoke } from "@tauri-apps/api/core";
 import { listen } from "@tauri-apps/api/event";
 import { onMounted, ref } from "vue";
