+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.ts",
   "scripts": {
-    "dev": "concurrently \"pnpm run dev:frontend\" \"pnpm run dev:backend\"",
+    "dev": "concurrently \"pnpm run dev:tauri\" \"pnpm run dev:backend\"",
     "dev:frontend": "pnpm --filter @monorepo/shared run dev",
     "dev:tauri": "pnpm  --filter @monorepo/shared run tauri",
     "dev:backend": "pnpm --filter @monorepo/shared run backend",
