+++ packages/shared/eslint-config/.eslintrc.json
@@ -44,6 +44,11 @@
     ],
     "unicorn/no-keyword-prefix": "off",
     "unicorn/prevent-abbreviations": "off",
-    "unicorn/prefer-module": "off"
+    "unicorn/prefer-module": "off",
+    "no-inline-comments": "error",
+    "no-warning-comments": [
+      "error",
+      { "terms": ["todo", "fixme", "xxx"], "location": "anywhere" }
+    ]
   }
 }
